+++ feed/index.html
@@ -14,7 +14,6 @@
     />
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"></script>
     <script src="/js/feedposts.js" type="module"></script>
-    <script src="/js/searchPosts.js" type="module"></script>
     <script src="/js/filterFeedPosts.js" type="module"></script>
     <script src="/js/utils/backToTopButton.js" type="module"></script>
     <script src="/js/createPost.js" type="module"></script>
