+++ js/filterFeedPosts.js
@@ -1,5 +1,7 @@
 import { allPostsResult } from "./feedposts.js";
 import { setUpHTML } from "./const.mjs";
+import { API_BASE_URL } from "./const.mjs";
+import { fetchAllUserPosts } from "./feedposts.js";
 
 document.getElementById("reactionPosts").addEventListener("click", async () => {
   try {
@@ -44,3 +46,39 @@ document.getElementById("datePosts").addEventListener("click", async () => {
     setUpHTML(post, postWallContainer);
   });
 });
+
+document.addEventListener("DOMContentLoaded", () => {
+  const searchBar = document.getElementById("searchBar");
+
+  searchBar.addEventListener("input", searchPosts);
+
+  const allPosts = `${API_BASE_URL}social/posts?&_comments=true&_author=true&_reactions=true&_count=true`;
+  fetchAllUserPosts(allPosts);
+});
+
+async function searchPosts() {
+  const searchInput = document.getElementById("searchBar").value.toLowerCase();
+  const postCards = document.querySelectorAll(".card");
+
+  postCards.forEach((postCard) => {
+    const postTitleElement = postCard.querySelector(".postTitle");
+    const postBodyElement = postCard.querySelector(".postBody");
+    const postAuthorElement = postCard.querySelector(".postAuthor");
+
+    if (postTitleElement && postBodyElement && postAuthorElement) {
+      const postTitle = postTitleElement.textContent.toLowerCase();
+      const postBody = postBodyElement.textContent.toLowerCase();
+      const postAuthor = postAuthorElement.textContent.toLowerCase();
+
+      if (
+        postTitle.includes(searchInput) ||
+        postBody.includes(searchInput) ||
+        postAuthor.includes(searchInput)
+      ) {
+        postCard.style.display = "block";
+      } else {
+        postCard.style.display = "none";
+      }
+    }
+  });
+}
